Fix remaining PR review issues
- make AssemblyScanner telemetry path tests OS-agnostic by building absolute paths with Path.Combine and Path.GetTempPath
- rename the false-positive benchmark output label from "Warmup findings" to "Baseline findings"
- make false-positive profile artifact filenames collision-resistant by including milliseconds in the timestamp
- propagate suspicious local-variable state using the effective MethodSignals instance in LocalVariableAnalyzer
- replace repeated ReflectionRule lookup in InstructionAnalyzer with the cached _reflectionRule field
- extract duplicate rule override reflection logic into Services/Helpers/RuleOverrideChecker and reuse it from InstructionAnalyzer and MethodScanner
