+++ Services/MethodScanner.cs
@@ -5,7 +5,6 @@ using MLVScan.Services.Helpers;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
 using System.ComponentModel;
-using System.Reflection;
 
 namespace MLVScan.Services
 {
@@ -54,7 +53,7 @@ namespace MLVScan.Services
 
             _rules = rules.ToArray();
             _stringLiteralRules = _rules
-                .Where(static rule => OverridesRuleMethod(rule, nameof(IScanRule.AnalyzeStringLiteral),
+                .Where(static rule => RuleOverrideChecker.OverridesRuleMethod(rule, nameof(IScanRule.AnalyzeStringLiteral),
                     typeof(string), typeof(MethodDefinition), typeof(int)))
                 .ToArray();
             _signalTracker = signalTracker ?? throw new ArgumentNullException(nameof(signalTracker));
@@ -306,18 +305,6 @@ namespace MLVScan.Services
             }
 
             return result;
-        }
-
-        private static bool OverridesRuleMethod(IScanRule rule, string methodName, params Type[] parameterTypes)
-        {
-            var method = rule.GetType().GetMethod(
-                methodName,
-                BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic,
-                binder: null,
-                types: parameterTypes,
-                modifiers: null);
-
-            return method != null && method.DeclaringType != typeof(IScanRule);
         }
     }
 }
